Move WalletApi to Core
