+++ app/server/src/main/scala/org/bitcoins/server/WalletRoutes.scala
@@ -4,12 +4,12 @@ import akka.actor.ActorSystem
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server._
 import org.bitcoins.commons.serializers.Picklers._
+import org.bitcoins.core.api.wallet.AnyHDWalletApi
 import org.bitcoins.core.currency._
 import org.bitcoins.core.protocol.transaction.Transaction
 import org.bitcoins.core.wallet.utxo.AddressLabelTagType
 import org.bitcoins.crypto.NetworkElement
 import org.bitcoins.node.Node
-import org.bitcoins.wallet.api.AnyHDWalletApi
 
 import scala.concurrent.Future
 import scala.util.{Failure, Success}
