+++ core/src/main/scala/org/bitcoins/core/api/keymanager/KeyManagerApi.scala
@@ -2,6 +2,7 @@ package org.bitcoins.core.api.keymanager
 
 import org.bitcoins.core.crypto.ExtPublicKey
 import org.bitcoins.core.hd.{HDAccount, HDPath}
+import org.bitcoins.core.wallet.keymanagement.KeyManagerParams
 import org.bitcoins.crypto.Sign
 
 import scala.util.Try
@@ -12,4 +13,5 @@ trait BIP39KeyManagerApi extends KeyManagerApi {
   def toSign(privKeyPath: HDPath): Sign
   def deriveXPub(account: HDAccount): Try[ExtPublicKey]
   def getRootXPub: ExtPublicKey
+  def kmParams: KeyManagerParams
 }
