+++ core/src/main/scala/org/bitcoins/core/api/wallet/SpvWalletApi.scala
@@ -1,4 +1,4 @@
-package org.bitcoins.wallet.api
+package org.bitcoins.core.api.wallet
 
 import org.bitcoins.core.bloom.BloomFilter
