+++ wallet-test/src/test/scala/org/bitcoins/wallet/ProcessTransactionTest.scala
@@ -1,10 +1,10 @@
 package org.bitcoins.wallet
 
+import org.bitcoins.core.api.wallet.WalletApi
 import org.bitcoins.core.currency._
 import org.bitcoins.testkit.Implicits._
 import org.bitcoins.testkit.core.gen.TransactionGenerators
 import org.bitcoins.testkit.wallet.BitcoinSWalletTest
-import org.bitcoins.wallet.api.WalletApi
 import org.scalatest.FutureOutcome
 import org.scalatest.compatible.Assertion
