+++ wallet/src/main/scala/org/bitcoins/wallet/internal/AccountHandling.scala
@@ -29,7 +29,7 @@ private[wallet] trait AccountHandling { self: Wallet =>
           s"purpose field and ${DEFAULT_HD_COIN.coinType.toInt} coin field"))
 
   /** @inheritdoc */
-  override protected[wallet] def getDefaultAccount(): Future[AccountDb] = {
+  override def getDefaultAccount(): Future[AccountDb] = {
     for {
       account <- accountDAO.read((DEFAULT_HD_COIN, 0))
     } yield {
@@ -45,7 +45,7 @@ private[wallet] trait AccountHandling { self: Wallet =>
   }
 
   /** @inheritdoc */
-  override protected[wallet] def getDefaultAccountForType(
+  override def getDefaultAccountForType(
       addressType: AddressType): Future[AccountDb] = {
     val hdCoin = addressType match {
       case Legacy       => HDCoin(HDPurposes.Legacy, DEFAULT_HD_COIN_TYPE)
